update message + add imu data analysis code

+++ firmware/PIFWRobot/Core/Src/application/imu/mpu9250/mpu9250.c
@@ -164,12 +164,12 @@ int mpu9250_get_accel_gyro(float *ax, float *ay, float *az, float *gx, float *gy
 		return -1;
 	}
 
-    *ax = (int16_t)((int16_t)buffer[0]  << 8) | buffer[1];	*ax /= g_lsb_per_g;
-    *ay = (int16_t)((int16_t)buffer[2]  << 8) | buffer[3];  *ay /= g_lsb_per_g;
-    *az = (int16_t)((int16_t)buffer[4]  << 8) | buffer[5];  *az /= g_lsb_per_g;
-    *gx = (int16_t)((int16_t)buffer[8]  << 8) | buffer[9];	*gx /= g_lsb_per_dpfs;
-    *gy = (int16_t)((int16_t)buffer[10] << 8) | buffer[11]; *gy /= g_lsb_per_dpfs;
-    *gz = (int16_t)((int16_t)buffer[12] << 8) | buffer[13]; *gz /= g_lsb_per_dpfs;
+    *ax = (float)((int16_t)buffer[0]  << 8 | buffer[1]);	*ax /= g_lsb_per_g;
+    *ay = (float)((int16_t)buffer[2]  << 8 | buffer[3]);  	*ay /= g_lsb_per_g;
+    *az = (float)((int16_t)buffer[4]  << 8 | buffer[5]);  	*az /= g_lsb_per_g;
+    *gx = (float)((int16_t)buffer[8]  << 8 | buffer[9]);	*gx /= g_lsb_per_dpfs;
+    *gy = (float)((int16_t)buffer[10] << 8 | buffer[11]); 	*gy /= g_lsb_per_dpfs;
+    *gz = (float)((int16_t)buffer[12] << 8 | buffer[13]); 	*gz /= g_lsb_per_dpfs;
 
 	return 0;
 }
